refactor: 하위 폴더에서 관리하도록 수정
