+++ hooks/src/lib/index.ts
@@ -1,6 +1,6 @@
-export { default as useCardNumbers } from './useCardNumbers';
-export { default as useCardCompany } from './useCardCompany';
-export { default as useCardCVC } from './useCardCVC';
-export { default as useCardExpirationDate } from './useCardExpirationDate';
-export { default as useCardOwner } from './useCardOwner';
-export { default as useCardPassword } from './useCardPassword';
+export { default as useCardNumbers } from './useCardNumbers/useCardNumbers';
+export { default as useCardCompany } from './useCardCompany/useCardCompany';
+export { default as useCardCVC } from './useCardCVC/useCardCVC';
+export { default as useCardExpirationDate } from './useCardExpirationDate/useCardExpirationDate';
+export { default as useCardOwner } from './useCardOwner/useCardOwner';
+export { default as useCardPassword } from './useCardPassword/useCardPassword';
