+++ hooks/src/lib/useCardExpirationDate/useMonthInput.ts
@@ -1,13 +1,21 @@
 import { ChangeEvent, FocusEvent, useState } from 'react';
-import useValidation, { ValidationType } from './useValidation';
-import Validation from './utils/validation';
+import useValidation, { ValidationType } from '../common/useValidation';
+import Validation from '../utils/validation';
 
-const CVC_LENGTH = 3;
+const EXPIRATION_DATE_LENGTH = 2;
+const MONTH = {
+  start: 1,
+  end: 12,
+} as const;
 
 const inputValidations: ValidationType<string>[] = [
   {
-    validate: (value) => Validation.isExactLength(CVC_LENGTH, value),
-    message: `${CVC_LENGTH}자리의 CVC번호를 입력해주세요.`,
+    validate: (value) => Validation.isExactLength(EXPIRATION_DATE_LENGTH, value),
+    message: `${EXPIRATION_DATE_LENGTH}자리 월(MM)을 입력해주세요. ex) 01`,
+  },
+  {
+    validate: (value) => Validation.isMonthInRange(Number(value)),
+    message: `${MONTH.start}부터 ${MONTH.end}사이의 숫자를 입력해주세요.`,
   },
 ];
 
@@ -18,13 +26,14 @@ const preventInputValidations: ValidationType<string>[] = [
   },
 ];
 
-const useCardCVC = (initialValue = '') => {
+const useMonthInput = (initialValue = '') => {
   const [value, setValue] = useState(initialValue);
   const { error, validateValue } = useValidation<string>();
   const isValid = value !== '' && !error.state;
 
   const onChangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
     const preventInputValidateResult = validateValue(e.target.value, preventInputValidations);
+
     if (!preventInputValidateResult) return;
 
     validateValue(e.target.value, inputValidations);
@@ -35,7 +44,7 @@ const useCardCVC = (initialValue = '') => {
     validateValue(e.target.value, inputValidations);
   };
 
-  return { value, isValid, error, onChange: onChangeHandler, onBlur: onBlurHandler };
+  return { value, error, isValid, onChange: onChangeHandler, onBlur: onBlurHandler };
 };
 
-export default useCardCVC;
+export default useMonthInput;
